backend/requirements.txt
# Core FastAPI and web server
fastapi==0.110.1
uvicorn==0.25.0

# HTTP client for TMDB API
httpx==0.27.0

# Data validation and parsing
pydantic==2.12.5

# Environment and configuration
python-dotenv==1.2.1

# CORS middleware
starlette==0.37.2